[client] Revamp how translations are handled
Translations have been moved to a folder `locales` and saved as `<language_code>.yaml`, so editing them is more comfortable than before and integration with the development environment is now possible too.

This is possible thanks to @intlify/unplugin-vue-i18n, which allows to compile the entire locales folder into a messages object to pass to the vue-i18n lib, and it is added as a vite plugin which has been added to the base config.